整理钩子代码；

+++ LeonDirectUI/Win32/Win32.cs
@@ -15,7 +15,7 @@ namespace LeonDirectUI.Win32
         public static extern IntPtr GetWindowLong(IntPtr hwnd, int nIndex);
 
         [DllImport("user32", CharSet = CharSet.Ansi, EntryPoint = "SetWindowLongA", ExactSpelling = true, SetLastError = true)]
-        public static extern int SetWindowLong(IntPtr hwnd, int nIndex, IntPtr dwNewinteger);
+        public static extern IntPtr SetWindowLong(IntPtr hwnd, int nIndex, IntPtr dwNewinteger);
 
         [DllImport("user32", CharSet = CharSet.Ansi, EntryPoint = "CallWindowProcA", ExactSpelling = true, SetLastError = true)]
         public static extern int CallWindowProc(IntPtr lpPrevWndFunc, IntPtr hwnd, int Msg, int wParam, int lParam);
